requirements.txt
kafka-python==2.0.1
pyyaml==5.3.1
psycopg2-binary==2.8.5
protobuf==3.12.2
requests==2.24.0
